db/db.go
package db

import (
	"context"

	"go.mongodb.org/mongo-driver/mongo"
	"go.mongodb.org/mongo-driver/mongo/options"
)

type db struct {
	*mongo.Client

	DBName string
}

type DB interface {
	CheckUpRepo
}

func NewDB(config Config) (DB, error) {
	clientOptions := options.Client().ApplyURI(config.ConnectionString)
	client, err := mongo.Connect(context.TODO(), clientOptions)
	if err != nil {
		return nil, err
	}

	if err := client.Ping(context.TODO(), nil); err != nil {
		return nil, err
	}

	return &db{Client: client, DBName: config.DBName}, nil
}
